+++ tests/coordinates/test_dense.py
@@ -87,6 +87,10 @@ class TestDenseCoordinate:
         assert DenseCoordinate([1, 2, 3]).equals(DenseCoordinate([1, 2, 3]))
         assert not DenseCoordinate([1, 2, 3]).equals(42)
 
+    def test_isdense(self):
+        with pytest.warns(FutureWarning, match="isdense"):
+            assert DenseCoordinate([1, 2, 3]).isdense() is True
+
     def test_get_indexer(self):
         assert DenseCoordinate([1, 2, 3])._get_indexer(2) == 1
         assert np.array_equiv(DenseCoordinate([1, 2, 3])._get_indexer([2, 3]), [1, 2])